# Convert endianess while copying in read/write into methods