Rather than first copying data from source to destination buffer and
then performing endianess adjustment, to the conversion while copying.
This means that each byte is accessed only once which (according to
benchmarks) speeds up read_xxx_into and write_xxx_into methods:

    | Benchmark                      | Before [ns/iter] | After [ns/iter] |
    |--------------------------------+------------------+-----------------|
    | slice_i64::read_big_endian     |  34,863  (±  30) | 23,656  (± 935) |
    | slice_i64::read_little_endian  |  15,518  (±  19) | 13,362  (± 405) |
    | slice_i64::write_big_endian    |  30,910  (± 109) | 23,123  (±  91) |
    | slice_i64::write_little_endian |  14,924  (±  21) | 13,209  (± 180) |
    |--------------------------------+------------------+-----------------|
    | slice_u16::read_big_endian     |   7,492  (± 343) |  3,788  (±  16) |
    | slice_u16::read_little_endian  |   3,366  (±   8) |  3,198  (±   3) |
    | slice_u16::write_big_endian    |   4,066  (±   7) |  4,497  (±   8) |
    | slice_u16::write_little_endian |   4,040  (± 946) |  3,193  (±   7) |
    |--------------------------------+------------------+-----------------|
    | slice_u64::read_big_endian     |  35,816  (± 251) | 23,259  (±  21) |
    | slice_u64::read_little_endian  |  15,506  (±  86) | 13,365  (±  81) |
    | slice_u64::write_big_endian    |  30,948  (±  63) | 23,102  (±  36) |
    | slice_u64::write_little_endian |  14,938  (±  17) | 13,158  (±  18) |

The benchmarks were done on AMD Ryzen 9 5900X 12-Core Processor.

I’m somewhat confused why little endian benchmark show improvements
but the results are reproducible.  My best guess is that it’s
compiler failing to optimise out `for v $dst.iter_mut() { nop(); }`
loops currently present.