@BurntSushi 2023-08-20
Hi, we spoke on reddit. I briefly skimmed this PR and it is touching a fair bit of code, including `unsafe` code. That makes it more difficult for me to review and my review latency is quite high. Is this the smallest change that can be made that satisfies what you're trying to do? Or can it be made smaller?