@mina86 2023-08-21
Done.  Note that I haven’t rerun the benchmarks but Godbolt confirms that the old code uses memcpy followed by a loop doing the conversion while the new code skips memcpy and does conversion while copying.

The unsafe code that previous PR had was a simplified version of `[T]:as_chunks` which is currently experimental.  I like it because it removes `.try_into().unwrap()` calls.  Generated code is the same either way.

Lastly, this PR has two separate commits.  Best to merge them separately.  I can submit the first commit as a separate PR if you prefer.

Lastly, while working on this, I’ve also noticed two unnecessary uses of unsafe and sent separate PRs for those.