@BurntSushi 2023-10-06
This PR is on crates.io in `byteorder 1.5.0`.